Model-based spatially variable gene (SVG) calling

Performs normalisation of spatial transcriptomics data using spatially-dependent spot- and gene- specific size factors.

Usage

SpaNormSVG(spe, verbose = TRUE)

# S4 method for class 'SpatialExperiment'
SpaNormSVG(spe, verbose = TRUE)

Arguments

spe

a SpatialExperiment or Seurat object, with the count data stored in 'counts' or 'data' assays respectively, and a SpaNorm model fit.

verbose

a logical, specifying whether to show update messages (default TRUE).

Value

a SpatialExperiment or Seurat object with F-statistics, false discovery rates (FDRs). For SpatialExperiment objects, these are stored in the rowData.

Details

SpaNorm SVG calling works by using the SpaNorm model fit for data normalisation to perform a likelihood ratio test (LRT). The model used for normalisation is considered to be the full model. A second nested model is fit without the splines representing biology. These nested models are then compared using a LRT to identify genes where the splines representing biology contain strong signal.

Examples

library(SpatialExperiment)
library(ggplot2)

data(HumanDLPFC)

HumanDLPFC = SpaNorm(HumanDLPFC, sample.p = 0.05, df.tps = 2, tol = 1e-2)
#> (1/2) Fitting SpaNorm model
#> 201 cells/spots sampled to fit model
#> iter:  1, estimating gene-wise dispersion
#> iter:  1, log-likelihood: -1149654.530080
#> iter:  1, fitting NB model
#> iter:  1, iter:  1, log-likelihood: -1149654.530080
#> iter:  1, iter:  2, log-likelihood: -817875.051427
#> iter:  1, iter:  3, log-likelihood: -730289.096272
#> iter:  1, iter:  4, log-likelihood: -715477.819485
#> iter:  1, iter:  5, log-likelihood: -713420.649743
#> iter:  1, iter:  6, log-likelihood: -713065.316554
#> iter:  1, iter:  7, log-likelihood: -712982.402574
#> iter:  1, iter:  8, log-likelihood: -712957.598604 (converged)
#> iter:  2, estimating gene-wise dispersion
#> iter:  2, log-likelihood: -712613.990872
#> iter:  2, fitting NB model
#> iter:  2, iter:  1, log-likelihood: -712613.990872
#> iter:  2, iter:  2, log-likelihood: -712457.521967
#> iter:  2, iter:  3, log-likelihood: -712448.597511 (converged)
#> iter:  3, log-likelihood: -712448.597511 (converged)
#> (2/2) Normalising data
HumanDLPFC = SpaNormSVG(HumanDLPFC)
#> (1/3) Retrieving SpaNorm model
#> (2/3) Fitting nested SpaNorm model
#> 201 cells/spots sampled to fit model
#> iter:  1, estimating gene-wise dispersion
#> iter:  1, log-likelihood: -1149654.530080
#> iter:  1, fitting NB model
#> iter:  1, iter:  1, log-likelihood: -1149654.530080
#> iter:  1, iter:  2, log-likelihood: -818186.913558
#> iter:  1, iter:  3, log-likelihood: -736604.099951
#> iter:  1, iter:  4, log-likelihood: -723768.349498
#> iter:  1, iter:  5, log-likelihood: -722449.976988
#> iter:  1, iter:  6, log-likelihood: -722432.852076 (converged)
#> iter:  2, estimating gene-wise dispersion
#> iter:  2, log-likelihood: -722418.533962
#> iter:  2, fitting NB model
#> iter:  2, iter:  1, log-likelihood: -722418.533962
#> iter:  2, iter:  1, log-likelihood: -722418.533962
#> iter:  2, iter:  1, log-likelihood: -722418.533962
#> iter:  2, iter:  2, log-likelihood: -722418.533962
#> iter:  2, iter:  2, log-likelihood: -722418.533962
#> iter:  2, iter:  2, log-likelihood: -722418.533962
#> iter:  2, iter:  3, log-likelihood: -722418.533962 (converged)
#> iter:  3, log-likelihood: -722418.533962 (converged)
#> (3/3) Finding SVGs
#> 1244 SVGs found (FDR < 0.05)
head(rowData(HumanDLPFC))
#> DataFrame with 6 rows and 5 columns
#>                   gene_name   gene_biotype     svg.F       svg.p     svg.fdr
#>                 <character>    <character> <numeric>   <numeric>   <numeric>
#> ENSG00000188976       NOC2L protein_coding   9.69062 8.48251e-08 6.90020e-07
#> ENSG00000188290        HES4 protein_coding   0.00000 1.00000e+00 1.00000e+00
#> ENSG00000187608       ISG15 protein_coding   0.00000 1.00000e+00 1.00000e+00
#> ENSG00000188157        AGRN protein_coding   4.60648 1.03606e-03 4.93809e-03
#> ENSG00000078808        SDF4 protein_coding   0.00000 1.00000e+00 1.00000e+00
#> ENSG00000176022     B3GALT6 protein_coding   0.00000 1.00000e+00 1.00000e+00